Fix main_process.js so that packaging function is wrapped in if statment

+++ main_process.js
@@ -1,12 +1,10 @@
 // @ts-nocheck
-// Basic init
 const { app, BrowserWindow, ipcMain, Menu, Dialog } = require('electron');
 const createMenuBar = require('./backend/menuBar.js');
 const { fork, exec } = require('child_process');
 const path = require('path');
 const fs = require('fs');
 
-// To avoid being garbage collected
 let mainWindow;
 
 app.on('ready', () => {
@@ -15,34 +13,36 @@ app.on('ready', () => {
   //Adding Menu Bar
   const menu = Menu.buildFromTemplate(createMenuBar(mainWindow, ResetDir, OpenDir));
   Menu.setApplicationMenu(menu);
-  const initialStartFlagFilePath = path.join(__dirname, 'electronUserData', 'initialStartup.txt');
-  if (!fs.existsSync(initialStartFlagFilePath)) {
-    let pathToExecutable;
-    if (process.platform === 'darwin') {
-      pathToExecutable = path.join(__dirname, '..', '..', 'MacOS', 'bundle-bee');
-    } else if (process.platform === 'win32') {
-      pathToExecutable = path.join(__dirname, '..', '..', 'MacOS', 'bundle-bee');
-    }
-    fs.writeFile(
-      initialStartFlagFilePath,
-      `command to run executable:
+  if (process.env.NODE_ENV === 'packaging') {
+    const initialStartFlagFilePath = path.join(__dirname, 'electronUserData', 'initialStartup.txt');
+    if (!fs.existsSync(initialStartFlagFilePath)) {
+      let pathToExecutable;
+      if (process.platform === 'darwin') {
+        pathToExecutable = path.join(__dirname, '..', '..', 'MacOS', 'bundle-bee');
+      } else if (process.platform === 'win32') {
+        pathToExecutable = path.join(__dirname, '..', '..', 'MacOS', 'bundle-bee');
+      }
+      fs.writeFile(
+        initialStartFlagFilePath,
+        `command to run executable:
     open ${pathToExecutable}
     dirname: ${__dirname}
     filename: ${__filename}
     initalstartflagfilepath: ${initialStartFlagFilePath}
     `
-    );
-    exec(`open ${pathToExecutable}`, err => {
+      );
+      exec(`open ${pathToExecutable}`, err => {
+        app.exit(0);
+      });
+    } else {
+      fs.unlink(initialStartFlagFilePath, err => {
+        if (err) console.log(err);
+      });
+    }
+    mainWindow.on('close', () => {
       app.exit(0);
     });
-  } else {
-    fs.unlink(initialStartFlagFilePath, err => {
-      if (err) console.log(err);
-    });
   }
-  mainWindow.on('close', () => {
-    app.exit(0);
-  });
 });
 
 ipcMain.on('ondragstart', (event, filePath) => {
@@ -178,4 +178,4 @@ function ResetDir() {
   app.exit(0);
 }
 
-function OpenDir(build) {}
+function OpenDir(build) { }
